moved data file to be consistent with other modules and updated import paths
